Move functions inside migration to prevent eventually breaking the breaking
