+++ source/databricks/package/datamigration/migration_scripts/202302280900_Update_basis_data_folder_structure.py
@@ -17,10 +17,6 @@ from azure.core.paging import ItemPaged
 from os import path
 from package.datamigration.migration_script_args import MigrationScriptArgs
 from package.datamigration.migration_error import MigrationError
-from package.datamigration.file_system_client_operations import (
-    is_directory_empty,
-    move_and_rename_folder,
-)
 
 
 def apply(args: MigrationScriptArgs) -> None:
@@ -99,3 +95,25 @@ def migrate_batch(
                     f"Directory '{grid_area_directory.name}' not removed because it is not empty"
                 )
             file_system_client.delete_directory(grid_area_directory.name)
+
+
+def is_directory_empty(file_system_client: FileSystemClient, path: str) -> bool:
+    sub_paths = file_system_client.get_paths(path)
+    # ItemPaged has no __len__ method, so we need to iterate in order to determine whether there are any items
+    for unused in sub_paths:
+        return False
+    return True
+
+
+def move_and_rename_folder(
+    file_system_client: FileSystemClient, source_path: str, target_path: str
+) -> None:
+    directory_client = file_system_client.get_directory_client(source_path)
+    if not directory_client.exists():
+        print(
+            f"Skipping migration ({__file__}). Source directory not found:{source_path}"
+        )
+        return
+
+    target_path = path.join(file_system_client.file_system_name, target_path)
+    directory_client.rename_directory(new_name=target_path)
